backend/requirements.txt
fastapi
python-multipart
uvicorn
sqlalchemy
psycopg2-binary
opencv-python
Pillow
dlib